+++ samples/Cpp/TestCppClient/Utils.cpp
@@ -1,6 +1,6 @@
 ﻿/* Copyright (C) 2019 Interactive Brokers LLC. All rights reserved. This code is subject to the terms
  * and conditions of the IB API Non-Commercial License or the IB API Commercial License, as applicable. */
-#include "StdAfx.h"
+#include "stdafx.h"
 
 #include "Utils.h"
 #include <iostream>
@@ -16,7 +16,7 @@ inline bool Utils::is_base64(std::uint8_t c) {
 }
 
 std::vector<std::uint8_t> Utils::base64_decode(std::string const& encoded_string) {
-    int in_len = encoded_string.size();
+    int in_len = (int)encoded_string.size();
     int i = 0;
     int j = 0;
    int in_ = 0;
@@ -27,7 +27,7 @@ std::vector<std::uint8_t> Utils::base64_decode(std::string const& encoded_string
         char_array_4[i++] = encoded_string[in_]; in_++;
         if (i ==4) {
             for (i = 0; i <4; i++)
-                char_array_4[i] = base64_chars.find(char_array_4[i]);
+                char_array_4[i] = (uint8_t)base64_chars.find(char_array_4[i]);
 
             char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
             char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
@@ -44,7 +44,7 @@ std::vector<std::uint8_t> Utils::base64_decode(std::string const& encoded_string
             char_array_4[j] = 0;
 
         for (j = 0; j <4; j++)
-        char_array_4[j] = base64_chars.find(char_array_4[j]);
+        char_array_4[j] = (uint8_t)base64_chars.find(char_array_4[j]);
 
         char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
         char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
