+++ source/CppClient/client/EClientSocket.cpp
@@ -1,7 +1,7 @@
 ﻿/* Copyright (C) 2019 Interactive Brokers LLC. All rights reserved. This code is subject to the terms
  * and conditions of the IB API Non-Commercial License or the IB API Commercial License, as applicable. */
 
-#include "StdAfx.h"
+#include "stdafx.h"
 
 
 #include "EPosixClientSocketPlatform.h"
@@ -181,7 +181,7 @@ void EClientSocket::encodeMsgLen(std::string& msg, unsigned offset) const
 
 	assert( sizeof(unsigned) == HEADER_LEN);
 	assert( msg.size() > offset + HEADER_LEN);
-	unsigned len = msg.size() - HEADER_LEN - offset;
+	unsigned len = (unsigned)(msg.size() - HEADER_LEN - offset);
 	if( len > MAX_MSG_LEN) {
 		m_pEWrapper->error( NO_VALID_ID, BAD_LENGTH.code(), BAD_LENGTH.msg());
 		return;
@@ -238,7 +238,7 @@ bool EClientSocket::isSocketOK() const
 	return ( m_fd >= 0);
 }
 
-int EClientSocket::fd() const
+intptr_t EClientSocket::fd() const
 {
 	return m_fd;
 }
@@ -248,7 +248,7 @@ int EClientSocket::receive(char* buf, size_t sz)
 	if( sz <= 0)
 		return 0;
 
-	int nResult = ::recv( m_fd, buf, sz, 0);
+	int nResult = ::recv( m_fd, buf, (int)sz, 0);
 
 	if( nResult == -1 && !handleSocketError()) {
 		return -1;
